+++ src/main/java/fr/utbm/repository/SessionDeFormationDao.java
@@ -13,7 +13,18 @@ public class SessionDeFormationDao {
      * @param sessionDeFormation : la formation à ajouter
      * @param session : la session courante Hibernate
      */
-    public static void ajouter(SessionDeFormation sessionDeFormation, Session session) {
+    public static void ajouter(SessionDeFormation sessionDeFormation, Session session) throws Exception {
+
+        if(sessionDeFormation == null
+                || sessionDeFormation.getCours() == null
+                || sessionDeFormation.getEmplacement() == null) {
+            throw new Exception("Erreur : session de formation incomplète, " +
+                    "l'élement ne peut pas être ajouté en base de données");
+        }
+
+        if(sessionDeFormation.getId() > 0) {
+            throw new Exception("Erreur : l'identifiant de la session de formation est déjà présent en base de données");
+        }
 
         session.beginTransaction();
         session.persist(sessionDeFormation);
@@ -25,7 +36,18 @@ public class SessionDeFormationDao {
      * @param sessionDeFormation : l'objet java session de formation à jour
      * @param session : session courante hibernate
      */
-    public static void sauvegarder(SessionDeFormation sessionDeFormation, Session session) {
+    public static void sauvegarder(SessionDeFormation sessionDeFormation, Session session) throws Exception {
+
+        if(sessionDeFormation == null
+                || sessionDeFormation.getCours() == null
+                || sessionDeFormation.getEmplacement() == null) {
+            throw new Exception("Erreur : session de formation incomplète, " +
+                    "l'élement ne peut pas être sauvegardé en base de données");
+        }
+
+        if(sessionDeFormation.getId() == 0) {
+            throw new Exception("Erreur : Id inconnu en base de données");
+        }
 
         session.beginTransaction();
         session.merge(sessionDeFormation);
@@ -37,7 +59,18 @@ public class SessionDeFormationDao {
      * @param sessionDeFormation : session de formation à effacer
      * @param session : session hibernate courante
      */
-    public static void delete(SessionDeFormation sessionDeFormation, Session session) {
+    public static void delete(SessionDeFormation sessionDeFormation, Session session) throws Exception {
+
+        if(sessionDeFormation == null
+                || sessionDeFormation.getCours() == null
+                || sessionDeFormation.getEmplacement() == null) {
+            throw new Exception("Erreur : session de formation incomplète, " +
+                    "l'élement ne peut pas être supprimé de la base de données");
+        }
+
+        if(sessionDeFormation.getId() == 0) {
+            throw new Exception("Erreur : Id inconnu en base de données");
+        }
 
         session.beginTransaction();
         session.delete(sessionDeFormation);
